+++ app/mission/bart/+mag/+app/+bart/VisualizationManager.m
@@ -18,7 +18,9 @@ classdef VisualizationManager < mag.app.manage.VisualizationManager
 
             supportedControls = [mag.app.control.Field(@mag.bart.view.Field), ...
                 mag.app.control.PSD(@mag.bart.view.Spectrogram), ...
-                mag.app.control.Spectrogram(@mag.bart.view.Spectrogram)];
+                mag.app.control.SignalAnalyzer(["Input1", "Input2"]), ...
+                mag.app.control.Spectrogram(@mag.bart.view.Spectrogram), ...
+                mag.app.control.WaveletAnalyzer(["Input1", "Input2"])];
 
             if ~isempty(model) && model.HasAnalysis
 
@@ -40,7 +42,14 @@ classdef VisualizationManager < mag.app.manage.VisualizationManager
         function figures = visualize(this, analysis)
 
             command = this.SelectedControl.getVisualizeCommand(analysis.Results);
-            figures = command.call();
+
+            if command.NArgOut > 0
+                figures = command.call();
+            else
+
+                command.call();
+                figures = matlab.ui.Figure.empty();
+            end
         end
     end
 end