+++ resources/release-notes.md
@@ -1,5 +1,6 @@
 # App
 
+- Add Signal Analyzer and Wavelet Analyzer as Bartington visualization options
 - Redesign mission selection window to support all enumerated missions (in `mag.meta.Mission`)
 - Redesign how view-controllers are selected for each mission
 - Reduce duplication in definition of Field, PSD and Spectrogram view-controllers